WIP refactor
Refactored quite a bit, to make it easier (for me) to understand, and
drastically simplify (we don't need a fully featured drag and drop
library). I was able to completely remove two classes: enumformat, of
which I found a note mentioning SHCreateStdEnumFmtEtc as an alternative
but isn't supported pre windows 2000 (not my problem), and
dropsourcenotify, which I just shoved into dropsource.

NOTE this is still not compiling, it's just gotten too late today